move force decoders to their own file
