+++ ocpmodels/models/sfarinet.py
@@ -14,6 +14,7 @@ from ocpmodels.models.base_model import BaseModel
 from ocpmodels.models.utils.pos_encodings import PositionalEncoding
 from ocpmodels.modules.phys_embeddings import PhysEmbedding
 from ocpmodels.modules.pooling import Graclus, Hierarchical_Pooling
+from ocpmodels.models.force_decoder import ForceDecoder
 
 NUM_CLUSTERS = 20
 NUM_POOLING_LAYERS = 1
@@ -230,61 +231,6 @@ class OutputBlock(nn.Module):
         out = scatter(h, batch, dim=0, reduce="add")
 
         return out
-
-
-class LambdaLayer(nn.Module):
-    def __init__(self, func):
-        super(LambdaLayer, self).__init__()
-        self.func = func
-
-    def forward(self, x):
-        return self.func(x)
-
-
-class ForceDecoder(nn.Module):
-    def __init__(self, type, input_channel, model_configs, act):
-        """
-        Decoder predicting a force scalar per atom
-
-        Args:
-            type (str): Type of force decoder to use
-            model_config (dict): Dictionary of config parameters for the decoder's model
-            act (callable): Activation function (NOT a module)
-
-        Raises:
-            ValueError: Unknown type of decoder
-        """
-        super().__init__()
-        self.type = type
-        self.act = act
-        assert type in model_configs, f"Unknown type of force decoder: `{type}`"
-        self.model_config = model_configs[type]
-        if self.type == "simple":
-            self.model = nn.Sequential(
-                Linear(
-                    input_channel,
-                    self.model_config["hidden_channels"],
-                ),
-                LambdaLayer(act),
-                Linear(self.model_config["hidden_channels"], 3),
-            )
-        else:
-            raise ValueError(f"Unknown force decoder type: `{self.type}`")
-
-        self.reset_parameters()
-
-    def reset_parameters(self):
-        for layer in self.model:
-            if hasattr(layer, "reset_parameters"):
-                layer.reset_parameters()
-            else:
-                if hasattr(layer, "weight"):
-                    nn.init.xavier_uniform_(layer.weight)
-                if hasattr(layer, "bias"):
-                    layer.bias.data.fill_(0)
-
-    def forward(self, h):
-        return self.model(h)
 
 
 @registry.register_model("sfarinet")
